Store both a ScriptPromise and ScriptPromiseResolver on OngoingNavigation
ScriptPromiseResolver assumes that its associated promise will be
returned, and resolving the promise will happen later. It is not designed
to resolve before the promise is returned, and clears its reference to
the promise on resolve. By immediately grabbing the promise from the
resolver, we can clean up several small messes where AppHistory had to
work around this limitation.

This also makes it possible to synchronously fire the navigaterror
event, per issue 1237515.

Finally, we can now reject appHistory.goTo() promises on frame detach.
This allows us to pass
external/wpt/app-history/navigate/goTo-detach-in-onnavigate.html,
which was timing out before.

Bug: 1237515, 1219750
Change-Id: I104a473f847d67c9251af18609bb8626ee594eb2