Convert beforeSelectorOnCodeElement.html to WPT tests
This test was originally added for WebKit's bug 11197 [1] [2] as a crash
test. It is not clear whether a pixel test was needed to check visual
rendering but the description of the test suggests it is also testing
the circle before and the square after.

This test was later tweaked several times with some font-related rules
to keep ensuring it matches the pixel expectation, leading for example
to the introduction of the non-standard value '-webkit-monospace' that
is intended to be removed [3] or of a workaround for bug [4] that is
actually fixed now.

This CL upstream the original test to WPT as a crash test. It also
converts the visual part of the testing to a reftest. However, because
Chromium does not render counter-style "square" as specified, the
test remains internal for now [5].

Bug: 1065468

[1] https://bugs.webkit.org/show_bug.cgi?id=11197
[2] https://source.chromium.org/chromium/chromium/src/+/2203391429cd40f45325c1d275209c84b5c0dc57
[3] https://crbug.com/1065468
[4] https://crbug.com/939823
[5] https://github.com/w3c/csswg-drafts/issues/6200

Change-Id: I70e489e6824c9afed1087b23ecee27c1bb861dc0